requirements.txt
# MCP SDK
mcp>=0.9.0

# GraphQL client libraries
gql>=3.5.0
graphql-core>=3.2.3
aiohttp>=3.9.0

# HTTP/SSE server
starlette>=0.35.0
uvicorn>=0.25.0
sse-starlette>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
pydantic>=2.5.0
